refactor: moved validation realted functionality to database folder
