+++ app.py
@@ -6,12 +6,12 @@ from flask import Flask
 from flask_restful import Api
 from dotenv import load_dotenv, find_dotenv
 from database.db import init_db
+from database.error_handler import bad_request, error_handler_blueprint
 from resources.quote import QuotesApi, QuoteApi
 from resources.author import AuthorsApi, AuthorApi
 from resources.scrape import ScrapeQuotesApi, ScrapeAuthorsApi, ScrapeStatus
 from utils.logger import init_logger
 from utils.celery_app import init_celery
-from utils.error_handler import bad_request, error_handler_blueprint
 
 load_dotenv(find_dotenv())
