+++ database/error_handler.py
